Move useEvaluationContext implementation to product-editor
