+++ packages/js/product-editor/src/utils/register-product-editor-block-type.ts
@@ -2,16 +2,41 @@
  * External dependencies
  */
 import { Block, BlockConfiguration } from '@wordpress/blocks';
-
-/**
- * External dependencies
- */
 import { registerWooBlockType } from '@woocommerce/block-templates';
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore No types for this exist yet.
+// eslint-disable-next-line @woocommerce/dependency-group
+import { useEntityId } from '@wordpress/core-data';
 
 interface BlockRepresentation< T extends Record< string, object > > {
 	name?: string;
 	metadata: BlockConfiguration< T >;
 	settings: Partial< BlockConfiguration< T > >;
+}
+
+function useEvaluationContext( context: any ) {
+	const { productType } = context;
+
+	const productId = useEntityId( 'postType', productType );
+
+	const getEvaluationContext = ( select: any ) => {
+		const editedProduct = select( 'core' ).getEditedEntityRecord(
+			'postType',
+			productType,
+			productId
+		);
+
+		return {
+			...context,
+			editedProduct,
+		};
+	};
+
+	console.log( 'hi there!' );
+
+	return {
+		getEvaluationContext,
+	};
 }
 
 export function registerProductEditorBlockType<
@@ -25,9 +50,12 @@ export function registerProductEditorBlockType<
 		usesContext: [ ...( metadata.usesContext || [] ), 'productType' ],
 	};
 
-	return registerWooBlockType( {
-		name,
-		metadata: augmentedMetadata,
-		settings,
-	} );
+	return registerWooBlockType(
+		{
+			name,
+			metadata: augmentedMetadata,
+			settings,
+		},
+		useEvaluationContext
+	);
 }
